Refactor config classes to its own module
