+++ commands/run/default.config
@@ -1,11 +1,4 @@
 # Run profiles
-[test]
-pipeline = dna-const
-profile = wgs,hg38_masked
-trio = false
-case = test_case
-default_panel = "OMIM-AUTO"
-
 [wgs-single]
 pipeline = dna-const
 profile = wgs,hg38_masked
@@ -57,7 +50,7 @@ default_panel = "arftlig-myeloisk-malignitet"
 
 [rna-test]
 pipeline = rna-const
-profile = ...
+profile = cmd
 sample_type = single
 samples = hg002-rna
