+++ commands/run/help_classes/config_classes.py
@@ -1,6 +1,5 @@
 from configparser import ConfigParser, SectionProxy
 from logging import Logger
-from pathlib import Path
 import sys
 from typing import Dict, List, Optional, Union
 
@@ -52,7 +51,7 @@ class RunProfileConfig:
     raw_config: Dict[str, str]
 
     pipeline: str
-    profile: str
+    profile_name: str
     # single, trio, tumor_pair
     sample_type: str
     samples: List[str]
@@ -65,7 +64,7 @@ class RunProfileConfig:
         self.pipeline = parse_mandatory_section_argument(
             run_profile_conf, profile_key, logger, "pipeline"
         )
-        self.profile = parse_mandatory_section_argument(
+        self.profile_name = parse_mandatory_section_argument(
             run_profile_conf, profile_key, logger, "profile"
         )
         self.sample_type = parse_mandatory_section_argument(
@@ -173,6 +172,17 @@ class RunSettingsConfig:
             self._parse_setting(logger, default_settings_key, pipeline_settings_key, "cluster")
         )
 
+    def get_items(self):
+        
+        combined_settings = {}
+        for key, val in self._default_settings.items():
+            combined_settings[key] = val
+        
+        for key, val in self._pipeline_settings.items():
+            combined_settings[key] = val
+
+        return combined_settings.items()
+
     def _parse_setting(
         self,
         logger: Logger,
@@ -211,19 +221,9 @@ class RunConfig:
     config_parser: ConfigParser
     profile: RunProfileConfig
     settings: RunSettingsConfig
-    samples: Dict[str, SampleConfig]
+    samples: Dict[str, SampleConfig] = {}
 
     run_type: str
-
-    # log_base_dir: str
-    # trace_base_dir: str
-    # work_base_dir: str
-
-    # base: Path
-    # baseline_repo: Path
-    # repo: Path
-    # datestamp: bool
-    # trio: bool
 
     def __init__(self, logger: Logger, config_parser: ConfigParser, run_type: str):
         self.config_parser = config_parser
@@ -245,7 +245,7 @@ class RunConfig:
             logger.error(f'Run type section "pipeline-default" needs to be specified')
             sys.exit(1)
 
-        self.run_settings_config = RunSettingsConfig(
+        self.settings = RunSettingsConfig(
             logger,
             self.config_parser["pipeline-default"],
             self.config_parser[pipeline_setting_key],
@@ -273,7 +273,7 @@ class RunConfig:
 
     def get_setting_entries(self) -> Dict[str, str]:
         key_vals = {}
-        for key, val in self.config_parser["settings"].items():
+        for key, val in self.settings.get_items():
             key_vals[key] = val
         return key_vals
 
@@ -282,109 +282,3 @@ class RunConfig:
         for key, val in self.config_parser[run_type].items():
             key_vals[key] = val
         return key_vals
-
-
-class Case:
-    def __init__(
-        self,
-        id: str,
-        clarity_pool_id: str,
-        clarity_sample_id: str,
-        sex: str,
-        type: str,
-        read1: str,
-        read2: str,
-        father: Optional[str] = None,
-        mother: Optional[str] = None,
-        phenotype: str = "healthy",
-    ):
-        self.id = id
-        self.clarity_pool_id = clarity_pool_id
-        self.clarity_sample_id = clarity_sample_id
-        self.sex = sex
-        self.type = type
-        self.read1 = read1
-        self.read2 = read2
-        self.father = father or "0"
-        self.mother = mother or "0"
-        self.phenotype = phenotype
-
-    def __getitem__(self, key: str) -> str:
-        return getattr(self, key)
-
-
-class CsvEntry:
-
-    headers = [
-        "clarity_sample_id",
-        "id",
-        "type",
-        "sex",
-        "assay",
-        "diagnosis",
-        "phenotype",
-        "group",
-        "father",
-        "mother",
-        "clarity_pool_id",
-        "platform",
-        "read1",
-        "read2",
-        "analysis",
-        "priority",
-    ]
-
-    case_headers = [
-        "clarity_sample_id",
-        "id",
-        "type",
-        "sex",
-        "phenotype",
-        "father",
-        "mother",
-        "read1",
-        "read2",
-    ]
-
-    def __init__(
-        self,
-        group: str,
-        cases: List[Case],
-        priority: Optional[str],
-        assay: str,
-        analysis: str,
-        diagnosis: str,
-    ):
-        self.cases = cases
-
-        self.assay = assay
-        self.group = group
-        self.clarity_pool_id = "NA"
-        self.diagnosis = diagnosis
-        self.platform = "illumina"
-        self.analysis = analysis
-        self.priority = priority or "grace-lowest"
-
-    def header_str(self) -> str:
-        return ",".join(self.headers)
-
-    def __getitem__(self, key: str) -> str:
-        return getattr(self, key)
-
-    def __str__(self) -> str:
-        rows: List[str] = []
-        for case in self.cases:
-            row: List[str] = []
-            for header in self.headers:
-                if header in self.case_headers:
-                    value = case[header]
-                else:
-                    value = self[header]
-                row.append(value.strip('"'))
-            rows.append(",".join(row))
-        return "\n".join(rows)
-
-    def write_to_file(self, out_path: str):
-        with open(out_path, "w") as out_fh:
-            print(self.header_str(), file=out_fh)
-            print(str(self), file=out_fh)
